+++ src/PaymentSlip.php
@@ -1320,11 +1320,9 @@ abstract class PaymentSlip
     /**
      * Get all elements of the slip
      *
-     * @param bool $fillZeroes Whether to return the code line filled with zeros or not.
      * @return array All elements with their lines and attributes.
-     * @todo Consider extracting the parameter as settable property, e.g. $fillWithZeros
      */
-    public function getAllElements($fillZeroes = true)
+    public function getAllElements()
     {
         $paymentSlipData = $this->paymentSlipData;