+++ CHANGELOG.md
@@ -6,8 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- Removed the parameters from the getAllElements (API breaking).
+  Added gettable/settable properties instead.
 
 ### Fixed
+- Removed breaking code from example
 - Make getDisplayXXXX methods dependent on the related data setting(s)
   This fixes a newly introduced issue where calling the getAllElements method
   would throw an exception, if there was one or more disabled element data.